Mechanical warning cleanup across 29 plugins: unnamed stub parameters, size_t index casts
AudioProcessor program stubs drop their unused parameter names (as Tremolo already did) and
int indices into std containers and raw arrays are cast to size_t at the use site. No
behaviour change. Rebuilt all 29 VST3 targets: 0 errors; the two targeted warning classes
fell from about 430 instances to 10. 88 warnings of other classes remain (float-conversion
52, shadow 12, implicit-float-conversion 9, unused-variable 5, plus the 10 above), left for
a later pass. pluginval spot checks on Multi Delay, Reverb and Compressor pass.

+++ HyperPrismReimagined/Source/AutoPan/AutoPanProcessor.cpp
@@ -273,18 +273,18 @@ void AutoPanProcessor::processAutoPan(juce::AudioBuffer<float>& buffer)
         float currentRightGain = smoothedRightGain.getNextValue();
         
         // Store original signals
-        float leftInput = leftData[sample];
-        float rightInput = rightData[sample];
+        float leftInput = leftData[static_cast<size_t>(sample)];
+        float rightInput = rightData[static_cast<size_t>(sample)];
         
         // Apply panning to combined mono signal
         float monoSignal = (leftInput + rightInput) * 0.5f;
         
-        leftData[sample] = monoSignal * currentLeftGain;
-        rightData[sample] = monoSignal * currentRightGain;
+        leftData[static_cast<size_t>(sample)] = monoSignal * currentLeftGain;
+        rightData[static_cast<size_t>(sample)] = monoSignal * currentRightGain;
         
         // Accumulate for metering
-        leftLevelSum += std::abs(leftData[sample]);
-        rightLevelSum += std::abs(rightData[sample]);
+        leftLevelSum += std::abs(leftData[static_cast<size_t>(sample)]);
+        rightLevelSum += std::abs(rightData[static_cast<size_t>(sample)]);
         lfoSum += std::abs(phasedLFO);
     }
